transfer group link from files/metadata to submissions
