--- datameta/alembic/versions/20210329_a314e39f2cc1.py
@@ -1,28 +1,0 @@
-"""backpopulate Group with Submissions
-
-Revision ID: a314e39f2cc1
-Revises: cfb9cf1b7f6c
-Create Date: 2021-03-29 14:55:31.166052
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a314e39f2cc1'
-down_revision = 'cfb9cf1b7f6c'
-branch_labels = None
-depends_on = None
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('submissions', sa.Column('group_id', sa.Integer(), nullable=False))
-    op.create_foreign_key(op.f('fk_submissions_group_id_groups'), 'submissions', 'groups', ['group_id'], ['id'])
-    # ### end Alembic commands ###
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f('fk_submissions_group_id_groups'), 'submissions', type_='foreignkey')
-    op.drop_column('submissions', 'group_id')
-    # ### end Alembic commands ###
