+++ datameta/alembic/versions/20210330_7b16156a6bca.py
@@ -1,0 +1,36 @@
+"""transfer group link from files/metadata to submissions
+
+Revision ID: 7b16156a6bca
+Revises: cfb9cf1b7f6c
+Create Date: 2021-03-30 08:27:53.310950
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '7b16156a6bca'
+down_revision = 'cfb9cf1b7f6c'
+branch_labels = None
+depends_on = None
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint('fk_files_group_id_groups', 'files', type_='foreignkey')
+    op.drop_column('files', 'group_id')
+    op.drop_constraint('fk_metadatasets_group_id_groups', 'metadatasets', type_='foreignkey')
+    op.drop_column('metadatasets', 'group_id')
+    op.add_column('submissions', sa.Column('group_id', sa.Integer(), nullable=False))
+    op.create_foreign_key(op.f('fk_submissions_group_id_groups'), 'submissions', 'groups', ['group_id'], ['id'])
+    # ### end Alembic commands ###
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(op.f('fk_submissions_group_id_groups'), 'submissions', type_='foreignkey')
+    op.drop_column('submissions', 'group_id')
+    op.add_column('metadatasets', sa.Column('group_id', sa.INTEGER(), autoincrement=False, nullable=False))
+    op.create_foreign_key('fk_metadatasets_group_id_groups', 'metadatasets', 'groups', ['group_id'], ['id'])
+    op.add_column('files', sa.Column('group_id', sa.INTEGER(), autoincrement=False, nullable=False))
+    op.create_foreign_key('fk_files_group_id_groups', 'files', 'groups', ['group_id'], ['id'])
+    # ### end Alembic commands ###
